requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.1.7
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-telegram-bot==13.4.1
pytz==2021.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tornado==6.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzlocal==2.1
zope.interface==5.3.0
